eebos: determine the case name from the file name parameter even if the deck was set externally
this makes the problem's caseName() method work as expected for
multiplexed simulators like meebos. The approach taken here has the
disadvantage that the EclDeckFileName must always be set, but since
there are currently no simulators which use eebos as a library, this
does not cause any issues at the moment whilst it fixes a user visible
message to stdout.