Reduce possibility to shoot oneself into the foot for copyright infringement
Since it is not clear what the term "Dune source" in DUNE's template
exception means (it can be interpreted as either "source code
verbatimly distributed by the DUNE project" or as "source code derived
from code distibuted by the DUNE project"), the template exception
also applies for modified files like matrixmarketspecializations.hh if
the second interpretation applies. Let's play it safe and simply
remove all functionality based on modified DUNE source code; in eWoms,
all of this is anyway either completely unused or not enabled without
active user intervention.

Note 1: This issue only bites when distibuting binaries because
tarballs containing non-GPL2 and GPL2 software does not count the
non-GPL2 parts as derivative work of the GPL2 ones. (If this wasn't
the case, Linux distributions shipping GPL and non-GPL software/data
on the same medium would not be possible. Since such distributions are
a frequent sight, shipping GPL2 and GPL3+ code in the same package is
almost certainly okay as long as the unpacked files do not form a
derived work of each other.)

Note 2: In contrast to what the DUNE project asserts on its website,
its template exception is *not* a clone of libstdc++'s: Newer versions
of libstdc++ use *much* more waterproof legalese, and even the
template exception of outdated libstdc++ versions feature a small but
crucial difference: The old libstdc++ template exception says "file"
while DUNE's version states "Dune source"!

Note 3: Since DUNE's template exception might not apply to modified
source files, eWoms needs to be build using unmodified upstream DUNE
files to be 100% on the safe side if these eWoms binaries ought to be
distributed. This is because eWoms is GPLv3+ licensed while DUNE is
GPLv2, so DUNE's template exception must apply in order for binaries
to be distributable legally. Be aware that this assessment primarily
affects future binary packages of eWoms for Linux distributions.